Finish the implement of q_insert_tail
To implement `q_insert_tail` and meet the needs of executing in constant
time, I add new field called `tail` to store the address of the last
element of the list.
By adding this field, we can simplily get the address of the last
element without iterating the whole list. And so we can ensure it can
execute in constant time.