Fix the reset size in prepare_inputs
In `prepare_inputs` function, if the class is randomly chosen to be `0`
then it means that it is fixed class and we must clear the input_data
relating to it to 0.

But the data size we want to clean is not just a 2-byte `uint16_t` size,
the chunk size is 16 byte. So I use `memset` to deal with it.