Import post-processing into dudect
Since the measured distribution tend to be positively skewed to larger
execution times, according to the topic of dudect.

The topic provide two way to deal with this situation:
  1. cropping data: use percentiles to crop the data that is likely to
     be abnormal.
  2. higher order pre-processing: try to use centered product to
     avoid abnormal result of testing constant time.

The implementation includes new files called percentile.[ch] to provide
computing percentiles of the measure data. And I try to revise
`update_statistics` function to provide t-test on cropped and second
order pre-processing data.